backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
pydantic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart